C++14 + use make_unique and make_shared to make shared pointers

+++ ixwebsocket/IXWebSocketTransport.cpp
@@ -538,7 +538,7 @@ namespace ix
 
         while (_txbuf.size())
         {
-            int ret = _socket->send((char*)&_txbuf[0], _txbuf.size());
+            ssize_t ret = _socket->send((char*)&_txbuf[0], _txbuf.size());
 
             if (ret < 0 && (_socket->getErrno() == EWOULDBLOCK || 
                             _socket->getErrno() == EAGAIN))
